swap_face.py
import cv2

def swap_face_faceswap(video_path, face_image_path, output_path):
    print("Processando troca de rosto com FaceFusion (simulação)...")
    cap = cv2.VideoCapture(video_path)
    fourcc = cv2.VideoWriter_fourcc(*'mp4v')
    out = cv2.VideoWriter(output_path, fourcc, 25.0, (int(cap.get(3)), int(cap.get(4))))
    while True:
        ret, frame = cap.read()
        if not ret:
            break
        out.write(frame)  # Simula o frame processado
    cap.release()
    out.release()
    print(f"Vídeo salvo em {output_path}")
